+++ tools/integrity/test/ValidatorSetConstruction.java
@@ -48,126 +48,6 @@ public class ValidatorSetConstruction {
     public final ExpectedException exception = ExpectedException.none();
 
     @Test
-    public void transitiveSubIsBuiltCorrectly() {
-        Sub semanticSub = new Sub();
-        SchemaConcept mockSchemaConcept0 = mock(SchemaConcept.class);
-        when(mockSchemaConcept0.label()).thenReturn(Label.of("z"));
-        SchemaConcept mockSchemaConcept1 = mock(SchemaConcept.class);
-        when(mockSchemaConcept1.label()).thenReturn(Label.of("a"));
-        SchemaConcept mockSchemaConcept2 = mock(SchemaConcept.class);
-        when(mockSchemaConcept2.label()).thenReturn(Label.of("b"));
-        SchemaConcept mockSchemaConcept3 = mock(SchemaConcept.class);
-        when(mockSchemaConcept3.label()).thenReturn(Label.of("c"));
-        SchemaConcept mockSchemaConcept4 = mock(SchemaConcept.class);
-        when(mockSchemaConcept4.label()).thenReturn(Label.of("d"));
-        SchemaConcept mockSchemaConcept5 = mock(SchemaConcept.class);
-        when(mockSchemaConcept5.label()).thenReturn(Label.of("entity"));
-        SchemaConcept mockSchemaConcept6 = mock(SchemaConcept.class);
-        when(mockSchemaConcept6.label()).thenReturn(Label.of("relation"));
-        SchemaConcept mockSchemaConcept7 = mock(SchemaConcept.class);
-        when(mockSchemaConcept7.label()).thenReturn(Label.of("attribute"));
-        SchemaConcept mockSchemaConcept8 = mock(SchemaConcept.class);
-        when(mockSchemaConcept8.label()).thenReturn(Label.of("thing"));
-
-        Type type0 = new Type(mockSchemaConcept0);
-        Type type1 = new Type(mockSchemaConcept1);
-        Type type2 = new Type(mockSchemaConcept2);
-        Type type3 = new Type(mockSchemaConcept3);
-        Type type4 = new Type(mockSchemaConcept4);
-        Type type5 = new Type(mockSchemaConcept5);
-        Type type6 = new Type(mockSchemaConcept6);
-        Type type7 = new Type(mockSchemaConcept7);
-        Type type8 = new Type(mockSchemaConcept8);
-
-        Pair<Type, Type> sub1 = new Pair<>(type0, type1);
-        Pair<Type, Type> sub2 = new Pair<>(type1, type2);
-        Pair<Type, Type> sub3 = new Pair<>(type2, type5);
-        Pair<Type, Type> sub4 = new Pair<>(type3, type6);
-        Pair<Type, Type> sub5 = new Pair<>(type4, type6);
-        // meta schema subtyping
-        Pair<Type, Type> sub6 = new Pair<>(type5, type8);
-        Pair<Type, Type> sub7 = new Pair<>(type6, type8);
-        Pair<Type, Type> sub8 = new Pair<>(type7, type8);
-
-        semanticSub.add(sub1);
-        semanticSub.add(sub2);
-        semanticSub.add(sub3);
-        semanticSub.add(sub4);
-        semanticSub.add(sub5);
-        semanticSub.add(sub6);
-        semanticSub.add(sub7);
-        semanticSub.add(sub8);
-
-        // this should not error
-        TransitiveSub transitiveSub = semanticSub.noIdentityTransitiveSub();
-        assertNotNull(transitiveSub);
-    }
-
-    @Test
-    public void transitiveSubThrowsOnLoop() {
-        Sub semanticSub = new Sub();
-        SchemaConcept mockSchemaConcept0 = mock(SchemaConcept.class);
-        when(mockSchemaConcept0.label()).thenReturn(Label.of("z"));
-        SchemaConcept mockSchemaConcept1 = mock(SchemaConcept.class);
-        when(mockSchemaConcept1.label()).thenReturn(Label.of("a"));
-        SchemaConcept mockSchemaConcept2 = mock(SchemaConcept.class);
-        when(mockSchemaConcept2.label()).thenReturn(Label.of("b"));
-        SchemaConcept mockSchemaConcept3 = mock(SchemaConcept.class);
-        when(mockSchemaConcept3.label()).thenReturn(Label.of("c"));
-        SchemaConcept mockSchemaConcept4 = mock(SchemaConcept.class);
-        when(mockSchemaConcept4.label()).thenReturn(Label.of("d"));
-        SchemaConcept mockSchemaConcept5 = mock(SchemaConcept.class);
-        when(mockSchemaConcept5.label()).thenReturn(Label.of("entity"));
-        SchemaConcept mockSchemaConcept6 = mock(SchemaConcept.class);
-        when(mockSchemaConcept6.label()).thenReturn(Label.of("relation"));
-        SchemaConcept mockSchemaConcept7 = mock(SchemaConcept.class);
-        when(mockSchemaConcept7.label()).thenReturn(Label.of("attribute"));
-        SchemaConcept mockSchemaConcept8 = mock(SchemaConcept.class);
-        when(mockSchemaConcept8.label()).thenReturn(Label.of("thing"));
-
-        Type type0 = new Type(mockSchemaConcept0);
-        Type type1 = new Type(mockSchemaConcept1);
-        Type type2 = new Type(mockSchemaConcept2);
-        Type type3 = new Type(mockSchemaConcept3);
-        Type type4 = new Type(mockSchemaConcept4);
-        Type type5 = new Type(mockSchemaConcept5);
-        Type type6 = new Type(mockSchemaConcept6);
-        Type type7 = new Type(mockSchemaConcept7);
-        Type type8 = new Type(mockSchemaConcept8);
-
-        Pair<Type, Type> sub1 = new Pair<>(type0, type1);
-        Pair<Type, Type> sub2 = new Pair<>(type1, type2);
-        Pair<Type, Type> sub3 = new Pair<>(type2, type5);
-        Pair<Type, Type> sub4 = new Pair<>(type3, type6);
-        Pair<Type, Type> sub5 = new Pair<>(type4, type6);
-        // meta schema subtyping
-//        Pair<Type, Type> sub6 = new Pair<>(type5, type8); // rewire this into a loop rather than to 'thing
-
-        Pair<Type, Type> sub7 = new Pair<>(type6, type8);
-        Pair<Type, Type> sub8 = new Pair<>(type7, type8);
-
-        // LOOP inducing
-        Pair<Type, Type> subLoop = new Pair<>(type5, type0);
-
-        semanticSub.add(sub1);
-        semanticSub.add(sub2);
-        semanticSub.add(sub3);
-        semanticSub.add(sub4);
-        semanticSub.add(sub5);
-//        semanticSub.add(sub6);
-        semanticSub.add(sub7);
-        semanticSub.add(sub8);
-        semanticSub.add(subLoop);
-
-        // this should not error
-        exception.expect(IntegrityException.class);
-        exception.expectMessage("is in a loop in the transitive closure of sub, implying a loop in the type hierarchy");
-        semanticSub.validate();
-
-    }
-
-
-    @Test
     public void entityRelationAttributeSetsAreBuiltCorrectly() {
         Sub semanticSub = new Sub();
         SchemaConcept mockSchemaConcept0 = mock(SchemaConcept.class);
